requirements.txt
beautifulsoup4>=4.13.0
certifi>=2025.6.0
charset-normalizer>=3.4.0
idna>=3.10
ping3>=4.0.8
pyasn1>=0.6.0
pysmb>=1.2.11
python-nmap>=0.7.1
PyYAML>=6.0.0
requests>=2.32.0
setuptools>=60.0.0
soupsieve>=2.7
tqdm>=4.67.0
typing_extensions>=4.14.0
urllib3>=2.4.0
